Move smearer to Core
